feat: add dropdown component (#2420)
Adds support for a `Dropdown` component, that can be used on its own or
as part of a `Field`.

```jsx
<Dropdown name="dropdown" value="option3">
  <DropdownOption value="option1">Option 1</DropdownOption>
  <DropdownOption value="option2">Option 2</DropdownOption>
  <DropdownOption value="option3">Option 3</DropdownOption>
</Dropdown>
```

Similarly to `Input`, it optionally uses `value` to determine the
initial value/preselected dropdown item. This can also be used to
control the input when re-rendering with `snap_updateInterface`. The
dropdown will also trigger `InputChangeEvent` similarly to an `Input`.

This PR mostly adds validation and types for this new component as well
as some handling for the form state. This will need a follow-up PR in
the extension to complete integration. This PR also adds support for
`selectInDropdown` to `snaps-jest` and uses this new component in the
Interactive UI example Snap.

Progresses https://github.com/MetaMask/MetaMask-planning/issues/1575

---------

Co-authored-by: MetaMask Bot <metamaskbot@users.noreply.github.com>