Change types to allow regular array types as children (#2470)
In #2466 we changed the types and validation of JSX components to use
`NonEmptyArray<Type>` as children. This causes issues however when using
utility functions like `.map()` and `.filter()`, since they return a
regular `Type[]` type. This is not great for devex, since it would
require type assertions everywhere.

I've changed the types to accept `Type[]`, though we technically still
validate `NonEmptyArray<Type>`.