BREAKING: Defer creation of offscreen document in `OffscreenExecutionService` (#2473)
Previously, the offscreen document needed to be created from the
`OffscreenExecutionService` because the extension did not do any
offscreen set up. Nowadays though, the offscreen document is used by the
extension for functionality other than Snaps, and the set up is handled
from the extension.

Instead, we the offscreen service now accepts a `Promise` which should
resolve when the offscreen document is ready. The service will wait for
the promise to resolve before continuing.