fix: Enforce stricter size limits for source code and auxiliary files (#2527)
Following the change in the MetaMask extension to use MV3, the Snaps
execution environment is now using extension message passing as part of
the JSON-RPC stack. This in turn means that messages sent between
MetaMask and the Snap must be smaller than 64 MB per message. Because of
this, we need to enforce sizing limits for the source code of the Snap
as well as any static auxiliary files that can be loaded at runtime.

This PR adds that validation on install, but also validates that any
returned value by `snap_getFile` is within these bounds, preventing
communication failure with the execution environment.

TODO: Add a warning or error in the CLI for these sizing limits?