fix: Add typed union validation (#2534)
Adds a `typedUnion` utility which strictly validates members of the
union based on a `type` field. This should reduce the noise when trying
to parse through Superstruct validation failures resulting in validating
large unions such as the JSX unions.

This doesn't solve the Superstruct union problem generally as we still
have many unions of unions, but it might make the problem easier to
solve.