feat: Add `SnapInsightsController` (#2555)
This PR adds `SnapInsightsController`, which is meant to handle fetching
a persistent response from transaction/signature insight Snaps. This
controller will temporarily store the insight response, letting the
client UI consume it wherever necessary.

The controller listens to events from `TransactionController` and
`SignatureController` to determine when new signatures or transactions
have entered the state. The events immediately trigger requests to any
available Snaps. The controller also manages cleaning up its own state
as well as the interface state when a confirmation finishes.

Most of the controller logic should mirror the implementation currently
found in the extension. With the exception that this controller now
populates the state as Snaps respond instead of waiting for all the
Snaps to load before triggering a state update.

Closes #2560