feat: Add RadioButton component (#2592)
Adds support for a `RadioGroup/Radio` component, that can be used on its
own or as part of a `Field`.

```jsx
<RadioGroup name="radiogroup" value="option3">
  <Radio value="option1">Option 1</Radio>
  <Radio value="option2">Option 2</Radio>
  <Radio value="option3">Option 3</Radio>
</RadioGroup>
```

Similarly to `Input`, it optionally uses value to determine the initial
value/preselected `Radio` item. This can also be used to control the
input when re-rendering with snap_updateInterface. The `RadioGroup` will
also trigger `InputChangeEvent` similarly to an `Input`.

This PR mostly adds validation and types for this new component as well
as some handling for the form state. This will need a follow-up PR in
the extension to complete integration. This PR also adds support for
`chooseFromRadioGroup` to `snaps-jest` and uses this new component in
the `Interactive UI example Snap`.

Related task: https://github.com/MetaMask/MetaMask-planning/issues/1571