feat: Add selector component (#2645)
Adds a new type of form component called `Selector`. It can be used as a
"rich" dropdown that allows displaying more than just text. Otherwise
functions very similarly to a dropdown from the Snaps platform
perspective.

It can be used as such:
```jsx
      <Selector name="selector" value="foo" title="Choose an option">
        <SelectorOption value="foo">
          <Card title="Foo" value="$1" />
        </SelectorOption>
        <SelectorOption value="bar">
          <Card title="Bar" value="$1" />
        </SelectorOption>
      </Selector>
```

Each client will have a different implementation of how the selector is
shown, but for reference the extension will show a modal that lets the
user pick between the options of the selector.

---------

Co-authored-by: MetaMask Bot <metamaskbot@users.noreply.github.com>