Extract simulation part of `snaps-jest` to separate package (#2727)
This extracts all simulation logic from `snaps-jest` to a new package
`snaps-simulation` (not to be confused with `snaps-simulator`). This
makes it possible to use the simulation framework outside of Jest, for
example to run a Snap in any Node.js application. It will also enable
removing the simulation framework used by `snaps-simulator` to reduce
code duplication, and improve maintainability.