Update `snap_notify` to allow JSX content (#2706)
Added the `title`, `detailedView` and `footerLink` properties to `inApp`
notifications. This allows snaps to provide expanded views with their
notifications. The `detailedView` property lets snaps return JSX content
with a limited subset of components. The `footerLink` property is
optional when providing an expanded view. This content is initially
being exposed to populate a modal, but in the future may populate an
entire page.