fix: Add missing dependency on execution environments to `snaps-simulation` (#2791)
When moving the simulation logic to `snaps-simulation`, we forgot to add
`snaps-execution-environments` as a direct dependency. This causes
errors at runtime when trying to install the latest version of
`snaps-jest`.

`snaps-execution-environments` is an optional peer dependency of
`snaps-controllers` . But is not optional for `snaps-simulation` as it
leverages the Node execution environments.