Add platform version field to manifest (#2803)
This adds a new field to the manifest, `platformVersion`. It is intended
to be set to the current version of the Snaps SDK, and if set, is
checked by the Snap controller to make sure the current supported
version by the Snaps Platform is equal or newer than the specified
`platformVersion`.

I've added a feature flag `rejectInvalidPlatformVersion` which, when
enabled, will throw an error upon installation if specified platform
version is too new. Otherwise it simply logs a warning in the console.