Re-introduce expanded view notifications via `snap_notify` (#2881)
This PR reverses the revert that was made to previous `snap_notify`
changes that enabled an expanded view for snap notifications (see
https://github.com/MetaMask/snaps/pull/2706). Additional adjustments
were made for including the content type of the interface being created
in the `snap_notify` call. These changes are being introduced
post-introduction of interface persistence in the
`SnapInterfaceController` (see
https://github.com/MetaMask/snaps/pull/2856), which without, the
expanded view feature was broken.