feat: Allow async initialization logic (#2918)
Allow async initialization logic for Snaps by altering the CLI config to
do direct assignment of the exports to `module.exports`. This lets Snap
developers use async initialization logic such as top-level await (which
has been re-enabled) and asynchronous instantiation of WASM modules
(which has been introduced in favor of sync). A minor change was also
made to the execution environment to await the exposed module after
evaluating it.

Snaps are still subject to a 60 second initialization timeout, so the
asynchronous logic must finish before that timer runs out.

Closes https://github.com/MetaMask/snaps/issues/2919
Closes https://github.com/MetaMask/snaps/issues/2359