feat: Automatically add `platformVersion` to manifest (#2938)
In https://github.com/MetaMask/snaps/pull/2803 we added support for
`platformVersion` in the manifest to indicate what version of the
platform the Snap was implemented for. We chose to not automatically add
this version to the manifest before releasing the initial changes. Since
those changes have now been released we can enable automatically adding
the field to the manifest.

Closes https://github.com/MetaMask/snaps/issues/2854