fix: Skip unnecesary provider initialization (#2967)
Since we only expose the `request` function from the provider and don't
allow listening to events or using `provider.chainId` etc we can
initialize the provider without making the `metamask_getProviderState`
request. This saves us a potential network request that may add overhead
to booting the Snap. Specifically the clients call `net_version` since
this property is required in the `MetaMaskInpageProvider`.

Closes https://github.com/MetaMask/snaps/issues/2968