Add script to update Chrome version (#2986)
I've created a script to simplify bumping Chrome and `chromedriver` a
little bit. With the script you can just run `yarn update-chrome`, which
fetches the latest version of `chromedriver`, and updates the version
and shasum in the `install-chrome` script.

Ideally we can run this automatically on Dependabot PRs, but Dependabot
seems to be broken right now.