Handle Webpack and Prettier breaking changes (#3103)
This updates the CLI and Webpack plugin to handle breaking changes
following the bump required for ESLint 9.

Unfortunately Prettier 3 makes it very complicated to load configuration
files with CJS, since it loads the ESM bundle using a top-level await,
which is not supported by Node.js without additional experimental
parameters. For now I've removed the functionality for custom Prettier
configs, but we can reconsider this in the future (i.e., through
specifying a Prettier path in the CLI config, or allowing to provide the
options directly).