Change Git hooks to be optional (#3211)
This changes Git hooks to be opt-in by running `yarn setup`. This is an
attempt to fix Dependabot, which [seems to be having issues
again](https://github.com/MetaMask/snaps/actions/runs/13781637142/job/38540795812).
This is just a guess at this point though, since the logs don't provide
much information about what's going on.