perf!: Use mnemonic seed for state persistence key derivation (#3217)
Bump key-tree to the latest version which includes a bunch of
performance improvements. One of these performance improvements is the
ability to use `.fromSeed`. This PR adds functionality to derive the
keys used for state encryption using the mnemonic seed instead of the
mnemonic phrase. This should save us more than half of the compute time
required for deriving keys used for state persistence.

This PR does not yet tackle `getEntropy` or `getBipEntropy` methods,
these will be supported in a follow-up.

Closes https://github.com/MetaMask/snaps/issues/3218

**Breaking changes**
- `SnapController` now expects a `getMnemonicSeed` hook that must return
the BIP39 seed for the user's primary mnemonic.