fix(send-flow-snap): add missing `displayAvatar` props, add tests, update `snaps-simulation` (#3312)
The send-flow example snap is only passing the `displayAvatar` prop in
the `to` scenario but not in the others. This can land you in a
situation where you paste a valid address in the to field, start to type
an amount in the amount field, then the avatar disappears because no
`displayAvatar` prop is being passed.

### Summary of changes

- Updated the send flow snap's `onUserInput` code to properly pass the
`displayAvatar` prop in all scenarios.
- Added tests for the snap.
- Added an SVG transformer so jest can handle the icons that are
imported in the snap.
- Updated the structure of the SVGs so it's consistent with what jest
expects.
- Updated `typeInField` in `snaps-simulation` so that it properly
attaches the `chainId` to the input value.