fix: Increase boot request queue limit (#3340)
Increase the limit of the `RequestQueue` to `100` pending requests. This
request queue is only used when the Snap has not yet booted, to detect
issues where the Snap is failing to boot. The current limit of `5` seems
too low to hit accidentally, therefore `100` was chosen which seems less
likely to hit with a real use-case.