feat: Add `snaps-sandbox` package and integrate into CLI (#3306)
This adds a new package, `@metamask/snaps-sandbox`, which is a tool to
interact with Snaps through a web app. It's a replacement for the Snaps
Simulator, but doesn't simulate any functionality. Instead, it calls
MetaMask through the usual JSON-RPC methods.

The sandbox can be used from the Snaps CLI by running `mm-snap sandbox`.