perf: Stop encoding messages in mobile (#3347)
The current approach for message passing in mobile encodes the JSON
payloads into a byte array before injecting into the WebView. This was
done by suggestion from @naugtur, but was known to be slow on mobile,
especially with larger messages.

More recently this caused performance degradations with Snaps that
MetaMask is actively trying to ship to production causing us to revisit
this. After discussion with @weizman we landed on the fact that simply
JSON stringifying should be safe enough and much more performant.