perf: Reduce unnecessary validation on responses (#3350)
Previously we were validating that responses were JSON, but we already
do that elsewhere and are guaranteed for the responses to be valid by
the time we get to `isValidResponse`. Additionally, we don't need an
entirely byte accurate length for the response, if the Snap goes beyond
the limit they simply won't receive a response from the Snap.