chore!: Build Snaps execution environments with Webpack (#3322)
This updates the Snaps execution environments build to use Webpack
instead of Browserify. The existing Browserify script was updated to
build with Webpack. In the future we may be able to use Webpack from the
CLI directly, but right now for compatibility with LavaMoat Node we need
to use a custom script.

I've added a new script `build:lavamoat:test` which builds the execution
environments for testing locally or in CI. The main difference with this
script is that it uses `tsconfig` paths, which are now disabled for the
production build. Doing so results in much better tree shaking, so a
smaller bundle.

## Breaking changes

- The `node-process` and `node-thread` bundles are now exported as
`@metamask/snaps-execution-environments/node-process` and
`@metamask/snaps-execution-environments/node-thread` respectively.