chore!: Remove web worker execution service and environment (#3371)
This is a redo of #3357, which was reverted in #3359 to avoid two
breaking changes in a row. Now that #3322 is merged, we can do this
again.

The original PR description can be found below.

---

This removes the web worker execution service and execution environment,
as they were unused. We can revert this PR in the future to add them
again, if needed.

## Breaking changes

### `@metamask/snaps-controllers`

- Remove web worker execution service.

### `@metamask/snaps-execution-environments`

- Remove web worker pool execution environment.
- Remove web worker executor execution environment.