fix: Properly catch cronjob errors during initialization (#3373)
This PR fixes two issues:

Firstly, we were not catching errors thrown during the "daily checkin"
of the `CronjobController`, this means that if one Snap failed we would
not complete the initialization of the controller and no cronjobs would
run.

Secondly, we were running all the cronjobs sequentially during "daily
check" which seems unintentional given that we don't need the result of
the cronjob invocation.