ci: Fix update workflow to use latest `yarn.lock` changes (#3379)
Currently we cache the `node_modules` folder and restore it from cache
in the steps after. This means that any changes to the `yarn.lock` file
made in the deduplication step are not used for generating LavaMoat
policies or building the examples, sometimes resulting in errors.

To fix it, I've added a check which checks if the `yarn.lock` was
updated. The other steps will use a high risk environment which skips
loading the `node_modules` cache if that's the case.