perf: Change request notification order (#3381)
Moved some of the ideas from https://github.com/MetaMask/snaps/pull/3356
to this PR which will be a simple to merge right away. The other
referenced PR needs some more work, but is eventually what we should do.

This PR let's Snaps start their outbound request before waiting for the
notification to land, this should be a small perf improvement to all
requests made from the Snap.