feat: Implement SIP-31 (#3394)
Introduces the `onClientRequest` handler as specified in SIP-31. This
handler can only be called by MetaMask and is available to all Snaps
without permission. It can be used for general purpose communication
that must not be exposed to dapps (as `onRpcRequest`).

This handler is enforced to only be callable by the `metamask` origin,
as part of this PR this same validation is applied to all handlers in
`CLIENT_ONLY_HANDLERS`.

Closes https://github.com/MetaMask/snaps/issues/3392

---------

Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>