feat: add background events example snap (#3404)
In our testing environment in the extension and probably so for mobile
as well, timers seem to have no delay so there is no reliable way to
execute a background event before the cronjob in the cronjobs examples
snap is executed, this makes it impossible to test both cronjobs and
background events in one snap.

**Summary of changes:**

- A new snap to test background events was created.
- The cronjob snap was updated to remove any background events logic.
- Test-snaps was updated to remove background events logic from the
cronjobs section.
- Background event logic was moved into its own section in test-snaps.