fix: Prevent scheduling background events less than 1 second in the future (#3414)
This PR fixes an issue with scheduling background events using
durations. Since we were truncating durations to remove milliseconds we
would effectively allow scheduling events less than 1 second into the
future.

To work around this, durations are now allowed to be a minimum of 1
second and are untruncated. The execution date will still be truncated
when the Snap requests it to prevent leakage of millisecond precision.

---------

Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>