feat: Detect unused permissions in Snaps CLI (#3335)
This adds unused permission detection for handlers to the Snaps CLI. It
will detect two cases:

- The Snap uses a certain permission, but does not export a handler for
it.
- The Snap exports a handler, but does not request permission for it.

Unfortunately due to how it works, it did require some refactors to the
CLI:

- The eval step in the build commands was removed in favour of the
Webpack plugin. Having a separate step made it complicated to support
things like watch mode, since that's handled by Webpack.
- Eval must run in order to detect unused permissions, so it's now run
in the `manifest` command as well.

---------

Co-authored-by: Frederik Bolding <frederik.bolding@gmail.com>