fix: Unwrap double-wrapped JSON-RPC errors (#3432)
A Snap may return a double-wrapped JSON-RPC error if it uses a JSON-RPC
method that throws. In that case we currently do not unwrap it properly
but instead return an error without a stack trace, this makes it
unreadable in `snaps-jest` and is just generally not what we want to do.
This PR unwraps this type of error to supply the stack trace.