fix: Recover from a crash more gracefully (#3440)
The current behavior for stopping Snaps (in case of crashes or idling)
was prone to race conditions when requests to the Snap were dispatched
around the time that the Snap crashed.

We saw this in production here:
https://github.com/MetaMask/metamask-extension/issues/33411

To prevent this, we can turn the `stopping` flag into a promise that we
can await in case of multiple calls to `stopSnap`, but also before
considering whether the Snap is started for a subsequent request.