fix: Fix unused permission detection for endowments with multiple handlers (#3452)
This fixes an edge case in the unused permission detection, for
endowments that can have multiple handlers. For example,
`endowment:lifecycle-hooks` enables two (optional) handlers, `onInstall`
and `onUpdate`. Previously the endowment would be detected as unused,
even if one of the two was exported.

After this change, either can be specified, and the endowment will be
considered used.