feat: Add WebSockets support (#3450)
Implements the rewritten SIP-20 to enable WebSocket support in Snaps.

WebSockets are supported via the `WebSocketService` which runs in the
client and handles open WebSocket connections for the Snap. The Snap is
notified of WebSocket events via `onWebSocketEvent` and is notified of
`open`, `close` and `message`. Messages can also be sent from the Snap
using `snap_sendWebSocketMessage`. Both string and binary payloads are
allowed.

All of the new RPC methods are gated on `endowment:network-access`,
which is also the permission required to expose `onWebSocketEvent`.