feat: Add `snap_trackError` method for error tracking through Sentry (#3498)
This adds a new RPC method `snap_trackError`, which can be used to track
errors in Sentry. It accepts a `name`, `message`, and `stack`, which
should be turned into a proper `Error` class again on the client end
before sending it to Sentry with `Sentry.captureException`.

To make the API a bit easier to use, I've added a `getJsonError` helper
function which takes an error (string, `Error` class, or other value)
and turns it into the properties expected by `snap_trackError`:

```ts
import { getJsonError } from '@metamask/snaps-sdk';

try {
  // ...
} catch (error) {
  await snap.request({
    method: 'snap_trackError',
    params: {
      error: getJsonError(error),
    },
  });
}
```

---------

Co-authored-by: MetaMask Bot <metamaskbot@users.noreply.github.com>