feat: Add feature flag to treat local Snaps as preinstalled (#3523)
This adds a new feature flag `forcePreinstalledSnaps`, which makes the
`SnapController` treat any local Snaps as preinstalled Snaps.

It's intended to only be enabled in local Flask builds, and should not
be enabled in any production builds. It can be used to simplify the
development of preinstalled Snaps by using a local Snap instead.

Related extension PR: MetaMask/metamask-extension#34022.

Closes #3517.