feat: Add `snap_trackError` and `snap_trackEvent` support to `snaps-jest` (#3546)
This adds support to `snaps-jest` for the `snap_trackError` and
`snap_trackEvent` methods. Internally in the simulation framework,
tracked errors and events are stored in state, and made available to the
response as `errors` and `events` properties. This can be combined with
the new `toTrackError` and `toTrackEvent` matchers to check if errors or
events are tracked.

I've updated the preinstalled example tests to use the new
functionality, and added an RPC method for tracking events as well.