feat: Allow eval in watch mode (#3553)
Evaluating the Snap bundle in watch mode was previously disabled because
it could result in, for example, permissions being removed from the
manifest immediately, because of some validator rules. In this PR, I've
updated the logic so only manifest errors are fixed immediately in watch
mode. Warnings are not fixed in watch mode, but are still fixed in build
or when running `mm-snap manifest --fix`.

Closes #3536.