fix: Ignore terminate calls to not executing Snaps (#3559)
When a Snap fails to start we call `terminateSnap`
(https://github.com/MetaMask/snaps/blob/main/packages/snaps-controllers/src/snaps/SnapController.ts#L3162),
but this may fail if the Snap has not fully started.

To prevent logging unnecessary errors to Sentry, this PR allows calls to
`terminateSnap` even if the Snap is not currently executing, if no job
is found the we just return early.