fix: Properly roll back `initialConnections` (#3618)
Our current rollback approach did not account for `initialConnections`,
which are handled separately from other permissions because they are
somewhat complicated to apply. If a Snap failed to update, we would not
correctly roll back the changes to the `initialConnections`. This PR
fixes that issue.

Fixes https://github.com/MetaMask/snaps/issues/3619