refactor!: Use hash private functions for updating internals (#3601)
The `updateSnap`, `processRequestedSnap` and `authorize` do not need to
be part of the public API and it makes maintaining them much simpler if
they aren't. This PR makes them all hash private and modifies tests to
use the `installSnaps` public API instead. This is also what is used in
production for all updating of Snaps.

Since these functions are guaranteed to be called only internally, this
PR also removes some dead code that was no longer needed.

This effort was moved out of the OTA branch to simplify review but does
contain **breaking changes**, namely the above functions no longer being
callable.