feat!: Allow updating preinstalled Snaps via the registry (#3616)
Allow updating preinstalled Snaps by updating their entry in the
registry. This is accomplished by searching the registry for updates
whenever `updateRegistry` is called and triggering `#updateSnap`. As
part of this PR the previously named `updateBlockedSnaps` has been
renamed to `updateRegistry` and a flag to allow automatic updates has
been added to `#updateSnap`.

When an update is automatic, the permissions requested by the Snap are
applied automatically similarly to how a preinstalled Snap is
initialized. Preinstalled Snaps can still be installed and updated via
the constructor. In case of any errors applying the update, the Snap is
rolled back to its previous state.

This behavior is put behind a feature flag that we can decide to enable
at a later date.

Closes https://github.com/MetaMask/snaps/issues/3593